access preferences via Preferences::ScopedStore
Objects including Preferable now always access preferences in the same
way through #preference_store, which returns an object responding to
fetch, []=, delete. This moves the responsibility of storage method
out of the object itself

ScopedStore is a proxy to the underlying cache store which adds a prefix
and suffix to the keys.