Remove order association from inventory_units
There is no need to store the order_id on an inventory_unit, as they are
tied to both a line_item in an order and a shipment in an order.
Instead, we can delegate order to the shipment.

This should hopefully allow some future simplification as they are more
normalized.