Automatically generate a spree api key when a user receives the admin role.
Given the recent change to only use api keys for admin api
authentication instead of the session cookie, an admin needs an api key
in order to do api things, and it does not receive one by default. It
seems kind to grant this to admin users when they become an admin, to
reduce the number of steps to give someone admin priveleges.

* Break out RoleUser HABTM into a first class model.
* Generate an api key for a new RoleUser's user if it doesn't exist and the role is an admin role
* Add primary key and timestamps to the spree_roles_users table